Move VerifyAccess annotation and Interceptor to better package
