+++ service/src/main/java/tds/exam/configuration/web/InterceptorConfiguration.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
-import tds.exam.utils.VerifyAccessInterceptor;
+import tds.exam.web.interceptors.VerifyAccessInterceptor;
 
 @Configuration
 public class InterceptorConfiguration extends WebMvcConfigurerAdapter {
