+++ service/src/main/java/tds/exam/web/annotations/VerifyAccess.java
@@ -1,4 +1,4 @@
-package tds.exam.utils;
+package tds.exam.web.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
